requirement.txt
boto==2.39.0
Django==1.8
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
djrill==2.0
filechunkio==1.6
gunicorn==19.4.5
MySQL-python==1.2.5
oauthlib==1.0.3
psycopg2==2.6.1
py==1.4.31
PyJWT==1.4.0
pytest==2.9.0
python-openid==2.2.5
python-social-auth==0.2.14
requests==2.9.1
requests-oauthlib==0.6.1
schedule==0.3.2
simplejson==3.8.2
six==1.10.0
wheel==0.26.0
xlrd==0.9.4
